requirements.txt
streamlit-fastapi-bridge